backend/requirements.txt
fastapi
uvicorn[standard]
gitpython
ollama
pydantic
python-dotenv
httpx
